// front/src/providers/api-client.generated/model/socialLoginRequest.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SocialLoginRequest { 
    /**
     * email of user
     */
    mail?: string;
    /**
     * password of user
     */
    password?: string;
    /**
     * firstName of user
     */
    firstName?: string;
    /**
     * lastName of user
     */
    lastName?: string;
    facebookUserId?: string;
    googleUserId?: string;
    twitterUserId?: string;
    photoUrl?: string;
}